Commented AreaMeter.z Refactored ConfirmationPopUp back to its generic self from before I modified it for tutorials to just call Quit() on confirmation. Now it takes a function from the request function which it calls if the confirmation is affirmative.
Refactored SpriteFader so it has black, white, and transparent fade to modes now.

UIButton recieved its first substantial refactor in years. Essentially instead
if disptaching a state variable within UIButtonEvent, now the UIButtonStateChange
event is dispatch under various state identifying String indentifiers. This
allows the UIMenuManager to identify the state of the button just by the name
the event is dispatched under.

UIMenuManager got a refactor alongside UIButton. OnUIButtonEvent function has
been refactored into multiple event responses instead of sorting by a if cases
and a state variable in the UIButtonEvent.

The preview project for this still needs to cover SpriteFader and HUDCreator.